# Conflicts:
#	app/src/main/java/com/okhome/awesomeapp/data/database/PhotoDao.kt
#	app/src/main/java/com/okhome/awesomeapp/views/detail/DetailFragment.kt
#	app/src/main/java/com/okhome/awesomeapp/views/photos/PhotosFragment.kt

+++ app/src/main/java/com/okhome/awesomeapp/data/database/PhotoDao.kt
@@ -6,7 +6,6 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import com.okhome.awesomeapp.module.database.PhotosEntity
-import com.okhome.awesomeapp.module.database.RemoteKeysEntity
 
 @Dao
 interface PhotoDao {
@@ -18,7 +17,7 @@ interface PhotoDao {
     fun getPhotos(): PagingSource<Int, PhotosEntity>
 
     @Query("SELECT * FROM ${PhotosEntity.TABLE_NAME} WHERE id = :id")
-    suspend fun getPhotoById(id: Int): PhotosEntity?
+    suspend fun getPhotoById(id: Long): PhotosEntity?
 
     @Query("DELETE FROM ${PhotosEntity.TABLE_NAME}")
     suspend fun clearPhotos()
